+++ src/Method/GetChatMember.php
@@ -1,0 +1,45 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vjik\TelegramBot\Api\Method;
+
+use Vjik\TelegramBot\Api\Request\HttpMethod;
+use Vjik\TelegramBot\Api\Request\TelegramRequestWithResultPreparingInterface;
+use Vjik\TelegramBot\Api\Type\ChatMember;
+use Vjik\TelegramBot\Api\Type\ChatMemberFactory;
+
+/**
+ * @see https://core.telegram.org/bots/api#getchatmember
+ */
+final readonly class GetChatMember implements TelegramRequestWithResultPreparingInterface
+{
+    public function __construct(
+        private int|string $chatId,
+        private int $userId,
+    ) {
+    }
+
+    public function getHttpMethod(): HttpMethod
+    {
+        return HttpMethod::GET;
+    }
+
+    public function getApiMethod(): string
+    {
+        return 'getChatMember';
+    }
+
+    public function getData(): array
+    {
+        return [
+            'chat_id' => $this->chatId,
+            'user_id' => $this->userId,
+        ];
+    }
+
+    public function prepareResult(mixed $result): ChatMember
+    {
+        return ChatMemberFactory::fromTelegramResult($result);
+    }
+}